MVC for login page
